+++ src/youwol/app/routers/python/__init__.py
@@ -1,0 +1,10 @@
+"""
+This module defines the [HTTP API router](https://fastapi.tiangolo.com/reference/apirouter/?h=apir) regarding
+  management of the in-browser python environment (through pyodide), it is served from the base HTTP URL:
+
+> **`/python`**
+
+"""
+
+# relative
+from .router import *